Version should increase, not reset to 1.0.0
A version check is performed by related plugins (WooCommerce https://github.com/algolia/algoliasearch-woocommerce) that rely on this plugin. The Algolia Search plugin for WooCommerce was also discontinued by Algolia at the same time as this plugin now supported by WebDevStudio.
When this project was forked by WebDevStudio, the version was reset from 2.11.3 to 1.0.0 .
Any reason? Please just set it to 2.11.3 or above! Easy fix and removes the incompatibility warnings.

The change in plugin folder name from `algoliasearch-wordpress` to `wp-search-with-algolia` was also problematic. I've also reverted that change in my site. Maybe we should just fork Algolia for Woocommerce as well. What are the plans for moving forward with this plugin? Any desire to maintain the fork of Algolia for Woocommerce?